+++ scripts/lichessTools.js
@@ -213,14 +213,22 @@
       };
     };
 
-    saveComment=(text, path)=>{
+    getNodeComment(node) {
+      const userId=this.getUserId();
+      const comment=(node?.comments||[]).find(c=>c?.by?.id==userId)?.text;
+      return comment;
+    }
+
+    saveComment=(text, path, chapterId)=>{
       const analysis=this.lichess?.analysis;
+      if (!chapterId) chapterId=analysis.study.currentChapter().id;
+      if (!path) path=analysis.path;
       analysis.study.makeChange('setComment', 
         {
-          ch: analysis.study.currentChapter().id,
-          path: path || analysis.path,
+          ch: chapterId,
+          path: path,
           text
-       });
+        });
     };
 
     isCommented(node) {
@@ -294,7 +302,13 @@
       return elem;
     }
 
+    assertPathSet(node) {
+      if (!node) return;
+      if (node.path===undefined) throw 'Path for node '+node.id+'( '+node.san+') not set!';
+    }
+
     getElementForNode(node) {
+      this.assertPathSet(node);
       const path=node.path||'';
       return this.getElementForPath(path);
     }
@@ -368,13 +382,12 @@
       return /\/following([\?#].*)?$/.test(this.global.location.href);
     };
 
-
-    findGlyphNode=(color,symbol)=>{
-      if (typeof symbol === 'string') symbol=[symbol];
+    findGlyphNode=(color,symbols)=>{
+      if (typeof symbols === 'string') symbols=[symbols];
       const analysis=this.lichess?.analysis;
       if (!analysis) return;
       const state=this.traverse();
-      const arr=[].concat.apply([],symbol.map(s=>state.glyphs[s]).filter(a=>!!a?.length));
+      const arr=[].concat.apply([],symbols.map(s=>state.glyphs[s]).filter(a=>!!a?.length));
       if (!arr.length) return;
       arr.sort((n1,n2)=>n1.nodeIndex-n2.nodeIndex);
       const index=analysis.node.nodeIndex;
@@ -382,14 +395,22 @@
       return arr.find(n=>n.ply%2===plyColor && n.nodeIndex>index)||arr.find(n=>n.ply%2===plyColor);
     };
 
-    jumpToGlyphSymbol=(color,symbol)=>{
+    jumpToGlyphSymbols=(symbols,side)=>{
       const analysis=this.lichess?.analysis;
       if (!analysis) return;
-      const node=this.findGlyphNode(color,symbol);
-      if (node) {
-        analysis.userJumpIfCan(node.path);
-        analysis.redraw();
+      let color='white';
+      if (['undefined','boolean'].includes(typeof side)) {
+        color=analysis.getOrientation();
+        if (side) {
+          color=color=='black'?'white':'black';
+        }
+      } else {
+        color=side;
       }
+      const node=this.findGlyphNode(color,symbols);
+      if (!node) return;
+      this.assertPathSet(node);
+      analysis.userJumpIfCan(node.path);
     };
 
     getPositionFromBoard=(el)=>{